requirements.txt
streamlit
python-dotenv
langchain
langgraph
langchain-openai
langchain-pinecone
pinecone-client
langchain-community
pypdf
typing-extensions
